+++ src/main/resources/cubicchunks.mixins.core_sided.bukkit.json
@@ -1,17 +1,17 @@
 {
     "required": false,
-    "package": "io.github.opencubicchunks.cubicchunks.core.asm.mixin.core",
+    "package": "io.github.opencubicchunks.cubicchunks.core.asm.mixin.core_sided.bukkit",
     "refmap": "cubicchunks.mixins.refmap.json",
     "compatibilityLevel": "JAVA_8",
     "minVersion": "0.7.10",
     "overwrites": {
         "conformVisibility": true
     },
-    "mixins": [],
-    "client": [],
-    "server": [
+    "mixins": [
         "common.MixinWorld_HeightLimits_Bukkit_Sided",
         "common.MixinPlayerList_Bukkit_Sided",
         "common.MixinChunk_Cubes_Bukkit_Sided"
-    ]
+    ],
+    "client": [],
+    "server": []
 }
